engine: MECE split — extract thread_context, script_resolver, action_runner
Per ADR-0013 (#480): engine.c mixed five concerns in one 377-line
file. Split into focused modules with single responsibilities:

  engine.c (181 lines)     — dispatcher: orchestrates the flow
  thread_context.c (89)    — per-thread state lifecycle (init/get/clear)
  script_resolver.c (78)   — find matching intercept_script in JSON config
  action_runner.c (103)    — iterate actions, dispatch via registry

Behavior is preserved exactly:
- Same call order in the dispatcher.
- Same log messages (including the misleading "failed to setup params"
  error that fires on the intentional FP-varargs bail — pre-existing,
  tracked separately).
- Same reentrance guard semantics.
- Same public API (retrace_engine_wrapper / retrace_engine_init).

The dispatcher now reads as a numbered list of the 14 steps, each
delegating to the module that owns it. Adding a new concern (e.g.
per-group log levels, va_list reconstruction for FP visibility)
becomes a new module rather than a growing engine.c.

Also removes the dead `struct WrapperSystemVFrame` declaration that
lived in engine.c but was never used there (each backend has its own
definition in arch_spec_bottom.c).

Verified locally on Apple Silicon:
- cmake --build clean, no new warnings in src/core.
- ctest 3/3 pass (integration + 2 unit suites).
- RETRACE_JSON_CONFIG=getuid-redirect.json → uid=0 (interception works).
- Non-FP printf("%d %s") round-trips with correct output + JSON logs.
