+++ src/core/CMakeLists.txt
@@ -9,7 +9,8 @@
 
 set(_core_sources
 	engine.c        funcs.c         actions.c      data_types.c
-	real_impls.c    logger.c        main.c         as_call_real.c)
+	real_impls.c    logger.c        main.c         as_call_real.c
+	thread_context.c script_resolver.c action_runner.c)
 
 # dlopen_guard.c is Linux-only: it exports dlopen/dlclose/dlsym/dlerror
 # as global symbols for LD_PRELOAD interposition. On macOS these symbols
